+++ cmd/tx.go
@@ -109,6 +109,16 @@ func createClientsCmd() *cobra.Command {
 		Args:    cobra.ExactArgs(1),
 		Example: strings.TrimSpace(fmt.Sprintf(`$ %s transact clients demo-path`, appName)),
 		RunE: func(cmd *cobra.Command, args []string) error {
+			allowUpdateAfterExpiry, err := cmd.Flags().GetBool(flagUpdateAfterExpiry)
+			if err != nil {
+				return err
+			}
+
+			allowUpdateAfterMisbehaviour, err := cmd.Flags().GetBool(flagUpdateAfterMisbehaviour)
+			if err != nil {
+				return err
+			}
+
 			c, src, dst, err := config.ChainsFromPath(args[0])
 			if err != nil {
 				return err
@@ -122,7 +132,8 @@ func createClientsCmd() *cobra.Command {
 				return err
 			}
 
-			modified, err := c[src].CreateClients(c[dst])
+			modified, err := c[src].CreateClients(c[dst], allowUpdateAfterExpiry,
+				allowUpdateAfterMisbehaviour)
 			if modified {
 				if err := overWriteConfig(config); err != nil {
 					return err
@@ -133,7 +144,7 @@ func createClientsCmd() *cobra.Command {
 		},
 	}
 
-	return cmd
+	return clientParameterFlags(cmd)
 }
 
 func updateClientsCmd() *cobra.Command {
@@ -219,6 +230,16 @@ $ %s tx conn demo-path --timeout 5s`,
 			appName, appName,
 		)),
 		RunE: func(cmd *cobra.Command, args []string) error {
+			allowUpdateAfterExpiry, err := cmd.Flags().GetBool(flagUpdateAfterExpiry)
+			if err != nil {
+				return err
+			}
+
+			allowUpdateAfterMisbehaviour, err := cmd.Flags().GetBool(flagUpdateAfterMisbehaviour)
+			if err != nil {
+				return err
+			}
+
 			c, src, dst, err := config.ChainsFromPath(args[0])
 			if err != nil {
 				return err
@@ -243,7 +264,8 @@ $ %s tx conn demo-path --timeout 5s`,
 			}
 
 			// ensure that the clients exist
-			modified, err := c[src].CreateClients(c[dst])
+			modified, err := c[src].CreateClients(c[dst], allowUpdateAfterExpiry,
+				allowUpdateAfterMisbehaviour)
 			if modified {
 				if err := overWriteConfig(config); err != nil {
 					return err
@@ -264,7 +286,7 @@ $ %s tx conn demo-path --timeout 5s`,
 		},
 	}
 
-	return retryFlag(timeoutFlag(cmd))
+	return clientParameterFlags(retryFlag(timeoutFlag(cmd)))
 }
 
 func closeChannelCmd() *cobra.Command {
@@ -321,6 +343,16 @@ $ %s tx connect demo-path`,
 			appName, appName, appName,
 		)),
 		RunE: func(cmd *cobra.Command, args []string) error {
+			allowUpdateAfterExpiry, err := cmd.Flags().GetBool(flagUpdateAfterExpiry)
+			if err != nil {
+				return err
+			}
+
+			allowUpdateAfterMisbehaviour, err := cmd.Flags().GetBool(flagUpdateAfterMisbehaviour)
+			if err != nil {
+				return err
+			}
+
 			c, src, dst, err := config.ChainsFromPath(args[0])
 			if err != nil {
 				return err
@@ -345,7 +377,8 @@ $ %s tx connect demo-path`,
 			}
 
 			// create clients if they aren't already created
-			modified, err := c[src].CreateClients(c[dst])
+			modified, err := c[src].CreateClients(c[dst], allowUpdateAfterExpiry,
+				allowUpdateAfterMisbehaviour)
 			if modified {
 				if err := overWriteConfig(config); err != nil {
 					return err
@@ -378,7 +411,7 @@ $ %s tx connect demo-path`,
 		},
 	}
 
-	return retryFlag(timeoutFlag(cmd))
+	return clientParameterFlags(retryFlag(timeoutFlag(cmd)))
 }
 
 func linkThenStartCmd() *cobra.Command {